src/requirements.txt
# Do not include azure-functions-worker as it may conflict with the Azure Functions platform

azure-functions
azure-storage-blob
azure-identity
azure-cognitiveservices-vision-computervision
azure-cognitiveservices-speech
